product katmanları oluşturuldu. builder() kullanımına geçildi lombok dependency si eklendi 2 tane constructor , getter ,setter ve builder annatasyonları eklendi

+++ src/main/java/com/hsgumussoy/javaodev2/repository/CategoryRepository.java
@@ -5,8 +5,8 @@ import com.hsgumussoy.javaodev2.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CategoryRepository extends JpaRepository<Category,Long> {
-    Category findCategoryByCategoryId(int id);
+    Category findCategoryByCategoryId(Long id);
 
-    Category deleteCategoryById(int id);
+    Category deleteCategoryById(Long id);
 
 }
